Fixed cell freeze problem when moving diagonally
The bug was caused by the fact that only the last overlapped cell was memorized instead of all of them. Because of this, what could happen was that _onCellLeft() could be called on a cell diagonally adjacent to the last overlapped one, freezing more cells than needed.

The bug has been fixed by making the grid subsystem store all currently overlapped cells.