EndFocus doesn't end interaction, EndInteraction public
Focus is required to start the interaction, interactions that already started do not require focus to keep going.

IInteractorInterface::EndInteraction() is public so that interactables can tell interactors that there is nothing more to do. This way, the interactor is removed from the interactionC's internal set at the correct time.