Crash fix: one inventory operation during pause
Previously, it was possible to queue both a drop and a usage operation on the inventory during pause. If the two operations were queued on the same object a crash occurred, because the item usage operation couldn't find the item anymore because of the drop.

Thanks to the command pattern on the tactical pause, now only one of the two operations can be queued during the pause, fixing the bug.