AI can now cast "Ice Wall"
Task is not UBTTask_BlackboardBase anymore. Maybe I'll need the node to be of this kind in the future (in case a part of the BT decides the skill to cast and the other executes it). For the time being, I decided to go for a UBTTaskNode, which is much simpler to manage.