AI casts Spells to change its Temperature (#51)
* Refactor: OnBeginSkillExecution() has skill param

* BTTask to cast spells

* Necessary components + RT-movement empty class

To test if the skill casting works properly, the UAIRotoTranslationMovement will be just an empty placeholder for the time being.
Later on, I'll have to implement a custom "Move To" BTTask that behaves differently depending on the current movement logic (that is updated by the skills).

* AI can now cast "Ice Wall"

Task is not UBTTask_BlackboardBase anymore. Maybe I'll need the node to be of this kind in the future (in case a part of the BT decides the skill to cast and the other executes it). For the time being, I decided to go for a UBTTaskNode, which is much simpler to manage.

* First task of custom "Move To" node

This node is needed for AI to make use of the Movement Command system I built time ago. I want this system because it is tightly linked to the skills and FSM systems.

* Custom move task: AI Move Request preparation

Taking inspiration from BTTask_MoveTo

* Custom move task: callback setup

* AI can cast Fire Globe and Cone of Cold

* Fix: Cone of Cold not castable while pathfinding

* Smooth AI rotation-only movement

* Option to not randomize in GetPosition task

* AI Widget shows mana and health

* AI Widget rename

* Rename of movement task

* Removing BT with old movement node

---------

Co-authored-by: Luca Pedrelli <Luca.Pedrelli@zuru.com>