Chaser Archetype (#53)
* Fire globe attaches to target

* Actors spawned by skill must implement interface

* Thermal Wind Actor, Impact Dmg Dealer

* Thermal Push refactored, now spawns thermal wind

The skill is working really bad for actors whose physics simulation isn't on, I'll probably need to find another solution based on overlaps instead of hits.

* First draft of Faction System

* AIControllers module, copy of existing controller

* Deletion of old enemy controller

* Copy of old controller renamed

* Chasers patrol the map

The patrolling logic needs to be moved to the character, otherwise I can't set the patrol per instance from the viewport

* Chaser section of BT (not working)

AEnemyChaserController::_onActorSensed() isn't getting called

* "MoveTo" points at enemy, perception sense added

AEnemyChaserController::_onActorSensed() now gets properly called, the patrol is interrputed in case the player is seen

* Patrol is now component on characters

* Thermal Push is now overlap-based

* Kinematic Impulse

* Old general purpose BB restored

* Custom movement command used by AI

The movement behavior isn't beautiful, but I'll keep it as is for the time being

* Temperature EQ-Test renamed

* EQ Test for relationship, not working on tester

Maybe a problem with stimuli source + AI perception? Try on real Chaser BT?

* Chaser BT performes enemies query regularly

* AI mov cmds broadcast changes to movement status

This is because we want the Blackboard to be aware of whether the controlled pawn is moving or not

* Moving status broadcast to AIController

* Chaser BT handles set target node properly

* Now chaser casts cone of cold

Well, the "unforseen consequences" arrived sooner than expected. Removing the delegate before stopping movement never warned the BT that the old movement mode was over, causing the movement node to never be called in conjunction with the !IsMoving BB-based decorator.

* Chaser cancels Cone of Cold if target is lost

* Chasers can cast Thermal Push

* Cone of Cold isn't interrupted for Thermal Push

* Task node details, tolerance on movement

* Set Movement Target nodes can be parametrized

---------

Co-authored-by: Luca Pedrelli <Luca.Pedrelli@zuru.com>