Make a new "Ice Shard" skill and edit Stabilizers to use it (#79)
* Ice Shard skill

* Remove check, caster might be dead after timer

* Function to check target validity

This avoids crashes where target checking on a dying actor skill target caused a check to trigger in UActorSkillTarget::GetTargetLocation()

* Only actual heat exchangers are employed

The deltaTemperature should be divided on the actors that actually exchanged heat with you, not all the ones detected by the simple collision

* Logs for debugging CoC spawn on player

* Other useful logs

* Bug: thermo actor overlap on spawn stops thermodyn

Cone of Cold (CoC) is cast => player controlled character is spawned => thermodynamics of character freezes. Same with trees. Why?
Because _counterOfChecksThisFrame overshoots _timesToBeCheckedThisFrame, causing _setCurrentTempAsNext() to never get called. Why?
Because spawning an actor means that the actor and all of its components will be ticked in the group it spawned, regardless of the tick group. The very next frame, the newly spawned actor will be ticked in the correct tick group.
This means that the following can happen:
- Character ticks, has been interacted every time already => resets counter
- CoC spawns, it ticks, and interacts with character => character's counter++
- New frame. CoC ticks before character => character's counter++
- Big problem: character's counter has been incresed 2 times by the same actor. Why is this a big problem?
- Because _timesToBeCheckedThisFrame is based on _possibleHeatExchangers.Num(), which is a TSet of unique elements => CoC is accounted only once

Fix: do not perform thermodynamics the frame that you spawn in the scene

* Thermodynamics overshoot logs removal

* Stabilizers' cast  ice shard on hot allies

* Crashes: managing dying actors

---------

Co-authored-by: Luca Pedrelli <Luca.Pedrelli@zuru.com>