Bugfix: count of detected non-ticking components
Say A spawns on frame X and interacts with B. They are each other's _heatExchangers. A cannot interact this frame, so B must also skip A. Problem is, we still need to pretend B's counter was updated by A, otherwise the increase in B's _timesToBeCheckedThisFrame wouldn't be compensated by B's counter.