Rework Thermodynamics to stop instancing air molecules (#90)
* ThermodynamicsCore module

* Preparations for Data-Oriented approach

* Grid initialization and debug

* Simple grid interaction

* Map extension and adjustments

* Grid to Grid and Body to Grid heat exchange

* Thermo presenter for actors

* Interaction between ThermoInteractors

To be tested properly once the game can execute without crashing if the old ThermoC is not on the controlled pawn.

* Fixes to interactor

* Proof that interaction between bodies works

* Initial substitution of thermoC

- Game does not crash when default pawn is new thermo character.
- The thermodynamics part of the old character has been re-made into the new one.
- Skills refactor to be completed (Thermal wind, ice wall). Fireball and Carbonize working.
- Animation fix.

* Fix heatmap out of bounds check

* Port IceWall to new thermodynamics

* Temperature dmg on new character

A new component is now required to perform death cleanup.

* Part of _die() logic moved out

Finish this so that AScalarFieldCharacter can finally be deleted

* DeathAnimation Component

* Impact damage handling rework

* Inventory and pickups on new character

* Chaser, Stabilizer, IceShard and Cone BPs

* Chaser as generic character works

* Redirection of AI-related UI

* Redirects cleanup

Redirectors can be safely removed once the affected assets, and the assets that use them, have been re-compiled and re-saved. This is why this commit has the thermometer BPs and enemy archetypes saved.

* Fixed walking animation of AICharacters

"Use acceleration for paths" must be set to true in the movement component in order for GetCurrentAcceleration() to properly work in the AnimBP.

* PlayerCharacter bUseControllerRotationYaw=false

The new PlayerCharacter wasn't properly rotating anymore, this fixes the issue.

* Bugfixing map

Makes it easy to debug 1 frame off problems

* Bugfix on targeting actors

If, during the targeting phase of a skill requiring actor targets, the player clicked without hitting any actor, a check was triggered. We don't want to trigger a check, but we also don't want the skill to consider the target valid.

* Bugfix: count of detected non-ticking components

Say A spawns on frame X and interacts with B. They are each other's _heatExchangers. A cannot interact this frame, so B must also skip A. Problem is, we still need to pretend B's counter was updated by A, otherwise the increase in B's _timesToBeCheckedThisFrame wouldn't be compensated by B's counter.

* Logs for 1 frame-off bugs

* Bugfix: interaction on unregistration

Unfortunately, actors can move during PrePhysics, potentially causing thermodynamic interaction to be non mutual. To fix this, interactions can now happen during _unregisterInteractor() if necessary.

* Fix comfort EQT to work with new thermo component

* Fix IceShard and FireGlobe to require new thermo C

* First half of Snowy Mountain terrain

* 2nd half of Snowy Mountain terrain

* Snowy Mountain trees

* New totem and replacements

* Default map back to Snowy Mountain

* Remove old thermo dependency from controller

* Fix ThermoPresenter logs

* Fix totems