Fix: multi-click targeting skill sets it available
Multi-clicking on a skill which needs targets, such as "Carbonize" or "Fire Ball", caused the icon of the screen to go from yellow (targeting) to green (available). This is because _resetWaitingSkill() was called by the container. This latter will now check whether the skill that is waiting isn't the same one we're re-trying to cast.