UI - Skills' status (#100)
* New Skill UI elements

* Creatiton of widgets c++ classes

* Making controllers responsible for initializations

SkillsContainerC must initialize before WidgetPresenterC for this latter to be able to retrieve and send skills data to the HUD.

* SkillInspection module

* Run of clang-format, LLVM 18.1.8

* Child Skill Icon Widgets added to Skill container

For some reason, they're still not visible on screen...

* Empty icons shown on screen

I forgot to specifiy the BP class deriving from USkillIconWidget.

* Remove useless UAbstractSkill::_onCooldown

Cooldown state can be checked via Timer Manager lol

* Cooldown starts after channeling, skill status

* Skill status is back to None after Cooldown

* Getters added to Skill Properties' Inspector

- Cooldown Time
- "Skill Status Changed" Delegate

* SkillIconWidget logic

Blueprint side is still missing

* Working WBP_SkillIcon

* Fix: skill waiting for targets can be aborted

* Thumbnail logic and textures

* Clickable icons (casting skills from UI)

* Don't restart cast of skill being executed

This is a weird "feature" I had which I find extremely weird + it let the player cast "thermal push" multiple times while paused without any effect being applied yet, (it's a skill that has a cast cost, no cast time, and has an effect during channelig).

* Fix thermal push multi cast during pause

See description of the previous commit, this is part 2 of the fix.

* Fix: multi-click targeting skill sets it available

Multi-clicking on a skill which needs targets, such as "Carbonize" or "Fire Ball", caused the icon of the screen to go from yellow (targeting) to green (available). This is because _resetWaitingSkill() was called by the container. This latter will now check whether the skill that is waiting isn't the same one we're re-trying to cast.